pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-index"
version = "0.1.0"
description = "Volatility surface analytics, macro-vol linkages, and regime detection for equity index options."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scipy>=1.11",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "plotly>=5.15",
    "PyYAML>=6.0",
    "pyarrow>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "mypy>=1.5",
    "ipykernel>=6.25",
    "jupyterlab>=4.0",
    "nbstripout>=0.6",
]
models = [
    "scikit-learn>=1.3",
    "statsmodels>=0.14",
    "hmmlearn>=0.3",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
